Move code for reading process availabilities into input/process/availability.rs
